# autogpt_platform/frontend/src/app/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoGPT Agent Server",
    "summary": "AutoGPT Agent Server",
    "description": "This server is used to execute agents that are created by the AutoGPT system.",
    "version": "0.1"
  },
  "paths": {
...
    "/api/blocks/{block_id}/execute": {
      "post": {
        "tags": ["v1", "blocks"],
        "summary": "Execute graph block",
        "operationId": "postV1Execute graph block",
        "parameters": [
          {
            "name": "block_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Block Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "type": "array", "items": {} },
                  "title": "Response Postv1Execute Graph Block"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/graphs/{graph_id}/execute/{graph_version}": {
      "post": {
        "tags": ["v1", "graphs"],
        "summary": "Execute graph agent",
        "operationId": "postV1Execute graph agent",
        "parameters": [
          {
            "name": "graph_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Graph Id" }
          },
          {
            "name": "graph_version",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Graph Version"
            }
          },
          {
            "name": "preset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Preset Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_postV1Execute_graph_agent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteGraphResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_postV1Execute_graph_agent": {
        "properties": {
          "inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputs"
          },
          "credentials_inputs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CredentialsMetaInput"
            },
            "type": "object",
            "title": "Credentials Inputs"
          }
        },
        "type": "object",
        "title": "Body_postV1Execute graph agent"
      },
...
      "CredentialsMetaInput": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "title": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Title"
          },
          "provider": { "$ref": "#/components/schemas/ProviderName" },
          "type": {
            "type": "string",
            "enum": ["api_key", "oauth2", "user_password", "host_scoped"],
            "title": "Type"
          }
        },
        "type": "object",
        "required": ["id", "provider", "type"],
        "title": "CredentialsMetaInput",
        "credentials_provider": [],
        "credentials_types": []
      },
...
      "ExecuteGraphResponse": {
        "properties": {
          "graph_exec_id": { "type": "string", "title": "Graph Exec Id" }
        },
        "type": "object",
        "required": ["graph_exec_id"],
        "title": "ExecuteGraphResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProviderName": {
        "type": "string",
        "enum": [
          "aiml_api",
          "anthropic",
          "apollo",
          "compass",
          "discord",
          "d_id",
          "e2b",
          "exa",
          "fal",
          "generic_webhook",
          "github",
          "google",
          "google_maps",
          "groq",
          "http",
          "hubspot",
          "ideogram",
          "jina",
          "linear",
          "llama_api",
          "medium",
          "mem0",
          "notion",
          "nvidia",
          "ollama",
          "openai",
          "openweathermap",
          "open_router",
          "pinecone",
          "reddit",
          "replicate",
          "revid",
          "screenshotone",
          "slant3d",
          "smartlead",
          "smtp",
          "twitter",
          "todoist",
          "unreal_speech",
          "zerobounce"
        ],
        "title": "ProviderName"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Postmark-Webhook-Token"
      }
    }
  }
}